make Utils tests standalone
